Broke Mapper back into its own file for better code searchability
